Simplifiy query.
I noticed that the only reason for the outer query was to solve the problem that the where condition filters out too much, because it requires the user_id to be a specific value, therefore completely negating the value of the outer join. Moving the check for the user_id into the on clause of the outer join actually fixed this as this explicitly does exactly that.